requirements.txt
ollama
google-generativeai
streamlit
groq
python-dotenv